requirements.txt
Django==1.11.20
django-filter==1.1.0
djangorestframework==3.9.2
pytz==2019.1
